Change in Linkedin layout

+++ Source/SSCore/src/main/java/com/realtech/socialsurvey/core/services/surveybuilder/impl/SurveyHandlerImpl.java
@@ -571,7 +571,9 @@ public class SurveyHandlerImpl implements SurveyHandler, InitializingBean {
 	@Transactional
 	public void deleteSurveyPreInitiationDetailsPermanently(SurveyPreInitiation surveyPreInitiation) {
 		LOG.info("Method deleteSurveyPreInitiationDetailsPermanently() started.");
-		surveyPreInitiationDao.delete(surveyPreInitiation);
+		surveyPreInitiation = surveyPreInitiationDao.findById(SurveyPreInitiation.class, surveyPreInitiation.getSurveyPreIntitiationId());
+		if(surveyPreInitiation != null)
+			surveyPreInitiationDao.delete(surveyPreInitiation);
 		LOG.info("Method deleteSurveyPreInitiationDetailsPermanently() finished.");
 	}
